Removed old prs stuff
