+++ app.conf.example
@@ -1,7 +1,5 @@
 [app]
 loglevel=debug
-keystore=hsm_api
-# keystore=hsm
 mtls_allow_override_cert=
 
 [database]
@@ -12,14 +10,6 @@ pool_size=5
 max_overflow=10
 pool_pre_ping=False
 pool_recycle=1800
-
-[auth]
-; When set, the server will use this certificate to authenticate itself to the client
-override_cert = ./secrets/prs-uzi.crt
-; Allowed curves for certificaets with EC keys
-allowed_curves = "curve25519 curve448 secp521r1 secp384r1 secp256r1"
-; Minimum RSA key size for certificates with RSA keys
-min_rsa_bitsize = 2048
 
 [uvicorn]
 swagger_enabled = True
@@ -33,56 +23,6 @@ use_ssl = False
 ssl_base_dir = secrets/ssl
 ssl_cert_file = server.cert
 ssl_key_file = server.key
-
-[rid]
-key_name = REK
-key_version = 1
-alg = AES-256-GCM
-key_renewal_at = 1000
-iv_prefix = RIVF
-; How old may a RID be before it is considered invalid for exchanging to a PDN/BSN via VAD?
-max_age_for_pdn_exchange_via_vad = 10
-; How old may a RID be before it is considered invalid for exchanging to a PDN/BSN via NZA/ZA?
-max_age_for_pdn_exchange_via_healthcare_provider = 3600
-
-[bpg]
-key_name = BPGK
-key_version = 1
-default_alg = HS256
-bp-gen-key-bpgk-1-alg=HS256
-
-[redis]
-host = redis
-port = 6379
-db = 0
-; Set these values to enable tls for redis connections
-ca_path =
-key_path =
-cert_path =
-
-
-[iv]
-; Path to IV counter
-path = secrets/iv.json
-; Block size to reserve. If the IV counter is below this value, a new block will be reserved and written to disks
-block_size = 100
-; Start value for the IV counter
-start_value = 9223372036854775807
-
-[hsm_keystore]
-library = /usr/local/lib/softhsm/libsofthsm2.so
-slot = 0
-slot_pin = 1234
-
-[hsm_api_keystore]
-; Url to the HSM api
-url = http://localhost:8000
-module = softhsm
-slot = SoftHSMLabel
-cert_path = secrets/hsm-api.crt
-
-[json_keystore]
-path=secrets/keystore.json
 
 [oprf]
 ; file with base64 server key for oprf evaluation (create with oprf_service.generate_key)
