+++ app.test.conf
@@ -1,7 +1,6 @@
 # Configuration settings for regular tests (locally, not CI)
 [app]
 loglevel=debug
-keystore=json
 mtls_allow_override_cert=
 
 [database]
@@ -12,13 +11,6 @@ pool_size=5
 max_overflow=10
 pool_pre_ping=False
 pool_recycle=1800
-
-[auth]
-override_cert=
-allowed_curves =
-min_rsa_bitsize = 2048
-allow_cert_list = auth_cert.json.example
-
 
 [uvicorn]
 swagger_enabled = False
@@ -31,48 +23,6 @@ use_ssl = False
 ssl_base_dir =
 ssl_cert_file =
 ssl_key_file =
-
-[rid]
-key_name = REK
-key_version = 1
-alg = AES-256-GCM
-key_renewal_at = 1000
-iv_prefix = IV
-max_age_for_pdn_exchange_via_vad = 10
-max_age_for_pdn_exchange_via_healthcare_provider = 3600
-
-[bpg]
-key_name = BPGK
-key_version = 1
-default_alg = HS256
-
-[redis]
-host = localhost
-port = 6379
-db = 0
-ca_path =
-key_path =
-cert_path =
-
-
-[iv]
-path = secrets/iv.json
-block_size = 100
-start_value = 9223372036854775807
-
-[hsm_keystore]
-slot = 1
-slot_pin = your_hsm_pin
-library = /path/to/your/pkcs11/library.so
-
-[hsm_api_keystore]
-url = https://your-hsm-api-endpoint
-module = your_module_name
-slot = your_slot_identifier
-cert_path = ""
-
-[json_keystore]
-path = secrets/keystore.json
 
 [oprf]
 ; file with base64 server key for oprf evaluation (create with oprf_service.generate_key)
